Rebasing with logins sort - updating logins store.
